refactor: fix broken imports after merge with master
