+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2019-02-06T14:59:33.323Z\n"
-"PO-Revision-Date: 2019-02-06T14:59:33.323Z\n"
+"POT-Creation-Date: 2019-02-06T16:00:53.261Z\n"
+"PO-Revision-Date: 2019-02-06T16:00:53.261Z\n"
 
 msgid "Cancel"
 msgstr ""
