+++ src/redux/reducers/index.js
@@ -4,7 +4,6 @@ import history from '../../utils/history'
 
 import feedback from './feedback'
 import standardReport from './standardReport'
-import periodTypes from './periodTypes'
 import dataSetReport from './dataSetReport'
 import reportPeriod from './reportPeriod'
 import organisationUnits from './organisationUnits'
@@ -13,7 +12,6 @@ export default combineReducers({
     router: connectRouter(history),
     feedback,
     standardReport,
-    periodTypes,
     dataSetReport,
     reportPeriod,
     organisationUnits,
